+++ src/lib/async/poly-iterable.ts
@@ -639,6 +639,55 @@ export default class PolyAsyncIterable<T> implements AsyncIterable<T> {
   }
 
   /**
+   * Returns the index of the first element for which `func(element)` returns `true`, or `-1` if it never does.
+   *
+   * @remarks
+   * `func` will be called on elements of this iteration until it returns `true`, and then not called again.
+   *
+   * Note that this method is rather useless given that iterables are single-use and have no indexing capabilities,
+   * but it's here for completion and consistency with `Array`.
+   *
+   * @param func - A boolean returning function called for elements of `this`
+   * @returns A promise the index of the first element of the iteration for which `func` returned `true`
+   */
+  async findIndex (func: AsyncIndexedPredicate<T>): Promise<number> {
+    asserts.isFunction(func)
+    let idx = 0
+    for await (const elem of this.#iterable) {
+      if (func(elem, idx)) {
+        return idx
+      }
+      idx++
+    }
+    return -1
+  }
+
+  /**
+   * Returns the index of the last element for which `func(element)` returns `true`, or `-1` if it never does.
+   *
+   * @remarks
+   * `func` will be called on *all* of this iteration, and the result will not be returned until the iteration ends.
+   *
+   * Note that this method is rather useless given that iterables are single-use and have no indexing capabilities,
+   * but it's here for completion and consistency with `Array`.
+   *
+   * @param func - A boolean returning function called for elements of `this`
+   * @returns A promise the index of the last element of the iteration for which `func` returned `true`
+   */
+  async findLastIndex (func: AsyncIndexedPredicate<T>): Promise<number> {
+    asserts.isFunction(func)
+    let foundIndex = -1
+    let idx = 0
+    for await (const elem of this.#iterable) {
+      if (func(elem, idx)) {
+        foundIndex = idx
+      }
+      idx++
+    }
+    return foundIndex
+  }
+
+  /**
    * Returns whether an element is present in this iteration.
    *
    * @remarks